refactor: remove permissions useless app, move Role, Permission models to staff app
