+++ backend/backend/apps/staff/admin.py
@@ -1,6 +1,16 @@
 from django.contrib import admin
 
-from backend.apps.staff.models import Staff
+from backend.apps.staff.models import Permission, Role, Staff
+
+
+@admin.register(Role)
+class RoleAdmin(admin.ModelAdmin):
+    pass
+
+
+@admin.register(Permission)
+class PermissionAdmin(admin.ModelAdmin):
+    pass
 
 
 @admin.register(Staff)
